requirements.txt
appdirs==1.4.4
APScheduler==3.6.3
asgiref==3.3.4
attrs==21.2.0
black==21.5b1
certifi==2020.12.5
click==8.0.0
Django==3.2.3
django-debug-toolbar==3.2.1
django-model-utils==4.1.1
Faker==8.1.4
iniconfig==1.1.1
mypy-extensions==0.4.3
packaging==20.9
pathspec==0.8.1
pluggy==0.13.1
py==1.10.0
pyparsing==2.4.7
pytest==6.2.4
pytest-django==4.2.0
python-dateutil==2.8.1
python-dotenv==0.17.1
python-telegram-bot==13.5
pytz==2021.1
regex==2021.4.4
six==1.16.0
sqlparse==0.4.1
text-unidecode==1.3
toml==0.10.2
tornado==6.1
tzlocal==2.1
